+++ config/terraform/aws/iam.tf
@@ -25,6 +25,8 @@ data "aws_iam_policy_document" "covidportal_secrets_manager" {
       aws_secretsmanager_secret.server_database_url.arn,
       aws_secretsmanager_secret_version.api_authorization.arn,
       aws_secretsmanager_secret_version.api_endpoint.arn,
+      aws_secretsmanager_secret_version.otk_sms_template_id_en.arn,
+      aws_secretsmanager_secret_version.otk_sms_template_id_fr.arn,
       aws_secretsmanager_secret_version.invitation_email_template_id_en.arn,
       aws_secretsmanager_secret_version.invitation_email_template_id_fr.arn,
       aws_secretsmanager_secret_version.confirmation_email_template_id_en.arn,