+++ profiles/static/js/script.js
@@ -19,3 +19,10 @@ if (form) {
   });
 }
 
+var radios = document.getElementsByName('otk-sent-next');
+radios.forEach(function(radio) {
+    radio.addEventListener('click', function(e) {
+        let value = document.querySelector('input[name="otk-sent-next"]:checked').value;
+        document.getElementById('otk_sms_sent_next').href = radio.value;
+    });
+});