+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- OTK sent via SMS using new html pages
+- OTK is temporarily cached in the session
+
 ## [1.15.1] - 2021-01-11
 
 ### Updated